Attack armour itp

+++ Assets/Scripts/Player/State/AliveState.cs
@@ -5,7 +5,7 @@ public class AliveState : PlayerState
         
     }   
     
-    public override void TakeDamage(int value)
+    public override void TakeDamage(float value)
     {
         _player.Stats.Health -= value;
         if(_player.Stats.Health <= 0) _player.SetState(new DeadState());
